[DF-6727] removed thumb from box2d and review fix.
Former-commit-id: e93cd1d9498687b0b852c14be629b3a1807e62d4

+++ Sources/Internal/Render/2D/TextLayout.h
@@ -94,7 +94,7 @@ public:
      * \brief Returns current cursor position
      * \return current cursor position
      */
-    const size_t Tell() const;
+    const uint32 Tell() const;
     
     /**
      * \brief Checks that text didn't finished yet
@@ -205,9 +205,9 @@ inline const TextLayout::WrapMode TextLayout::GetWrapMode() const
     return wrapMode;
 }
 
-inline const size_t TextLayout::Tell() const
+inline const uint32 TextLayout::Tell() const
 {
-    return fromPos;
+    return static_cast<uint32>(fromPos);
 }
